Разработка на Python/less_12/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyyaml==6.0.1
